Installed react-route-dom and wrapped App.js class with { BrowserRouter }, and moved all the components to Containers except Post. also created Posts new container
