+++ README.md
@@ -16,6 +16,13 @@ Axios.Delete(url) -> To Delete Data
 ---
 Added default URL to index.js and removed base URL from all the components.
 
+Branch: 6RouteConfig
+### `npm install --save react-router-dom`
+
+Now import "{ BrowserRouter }" from 'react-router-dom' and wrap out App.js component to enable routing mechanism all over the app.
+
+Put components into "Container" folder which needs to be routed. Here, Moving FullPost, NewPost from components to Container and also created another folder named "Posts".
+
 ## Available Scripts
 
 In the project directory, you can run:
