+++ src/containers/FullPost/FullPost.js
