+++ src/containers/NewPost/NewPost.css
